CSP-2354 Refactor: move GetAllStandardsQuery to dedicated namespace
