+++ src/SFA.DAS.Roatp.Application.UnitTests/Standards/Queries/GetAllStandards/GetAllStandardsQueryHandlerTests.cs
@@ -6,12 +6,12 @@ using FluentAssertions;
 using Microsoft.Extensions.Logging;
 using Moq;
 using NUnit.Framework;
-using SFA.DAS.Roatp.Application.Standards.Queries;
+using SFA.DAS.Roatp.Application.Standards.Queries.GetAllStandards;
 using SFA.DAS.Roatp.Domain.Entities;
 using SFA.DAS.Roatp.Domain.Interfaces;
 using SFA.DAS.Roatp.Domain.Models;
 
-namespace SFA.DAS.Roatp.Application.UnitTests.Standards.Queries
+namespace SFA.DAS.Roatp.Application.UnitTests.Standards.Queries.GetAllStandards
 {
     [TestFixture]
     public class GetAllStandardsQueryHandlerTests
