+++ src/SFA.DAS.Roatp.Application/Standards/Queries/GetAllStandards/GetAllStandardsQuery.cs
@@ -1,7 +1,7 @@
 ﻿using MediatR;
 using SFA.DAS.Roatp.Domain.Models;
 
-namespace SFA.DAS.Roatp.Application.Standards.Queries
+namespace SFA.DAS.Roatp.Application.Standards.Queries.GetAllStandards
 {
 
     public class GetAllStandardsQuery : IRequest<GetAllStandardsQueryResult>
